Create Sphinx project for documentation
Run the following commands to create a Sphinx documentation project:

    mkdir docs
    cd docs
    sphinx-quickstart -q -p MyHN -a "Sunaina Pai" --ext-autodoc

The `--ext-autodoc` option enables the `sphinx.ext.autodoc` extension in
`conf.py` of the Sphinx project. This will be needed later to process
`automodule` directive in `.rst` files to automatically include module
documentation from module docstrings.

In this commit, no files have been modified after creating the Sphinx
project, so that this commit can be used as a reference for what a
default Sphinx project looks like.

The Sphinx project files will be modified later to include actual
documentation of MyHN.