Fix tests to not patch open() and os.path.exists()
The `test_main_read_cache()` method fails in Travis CI builds with the
following error:

```
======================================================================
ERROR: test_main_read_cache (test.test_myhn.MyHNTest)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/opt/python/3.7.1/lib/python3.7/unittest/mock.py", line
  1191, in patched
    return func(*args, **keywargs)
  File "/home/travis/build/sunainapai/myhn/test/test_myhn.py", line
  90, in test_main_read_cache
    myhn.main()
  File "/home/travis/build/sunainapai/myhn/myhn.py", line 168, in main
    args = _parse_cli()
  File "/home/travis/build/sunainapai/myhn/myhn.py", line 55, in _parse_cli
    parser = argparse.ArgumentParser()
  File "/opt/python/3.7.1/lib/python3.7/argparse.py", line 1652, in __init__
    self._positionals = add_group(_('positional arguments'))
  File "/opt/python/3.7.1/lib/python3.7/gettext.py", line 625, in gettext
    return dgettext(_current_domain, message)
  File "/opt/python/3.7.1/lib/python3.7/gettext.py", line 588, in dgettext
    codeset=_localecodesets.get(domain))
  File "/opt/python/3.7.1/lib/python3.7/gettext.py", line 533, in translation
    t = _translations.setdefault(key, class_(fp))
  File "/opt/python/3.7.1/lib/python3.7/gettext.py", line 260, in __init__
    self._parse(fp)
  File "/opt/python/3.7.1/lib/python3.7/gettext.py", line 353, in _parse
    magic = unpack('<I', buf[:4])[0]
TypeError: a bytes-like object is required, not 'str'

----------------------------------------------------------------------
Ran 8 tests in 0.029s

FAILED (errors=1)
Makefile:20: recipe for target 'test' failed
make: *** [test] Error 1
```

This error occurs due to patching of `builtins.open` and
`os.path.exists` with mocks. When we call `argparse.ArgumentParser()`,
it leads to a call to `gettext.translation()`. This in turn calls
`gettext.find()` which performs `os.path.exists(mofile)` which returns
`True` due to patching of `os.path.exists`. As a result, it attempts to
open `mofile` in binary mode with the `open(mofile, 'rb')` call and then
tries to parse the binary data that it expects to read from `mofile`.
Since `open` is patched such that attempting to read the mock file
handle always returns `str` object with a value of `'{}'`, not a `bytes`
object as expected by the `gettext` module, it fails with the above
error.